+++ CHANGELOG.md
@@ -255,3 +255,4 @@
   - Previously, Claws II would say it needs Claws II, rather than needing Claws I. That's fixed now
 - The following Genes can now be held by any mob, rather than only players: Haste, More Hearts II, Night Vision
 - Removed classes for several Data Components, because they really weren't needed. This shouldn't affect anything
+- Moved several Data Component keys to snake_case from camelCase (for example, `dnaPoints` is now `dna_points`). This should be a non-breaking change